More robust SQL type handling
This commit does a number of things. First of all it fixes a bug where
dropping a default value was resulting into a SQL parse error. This is
because when dropping a constraint we don't have to specify the default
value, and therefore the syntax for SET/DROP needed to be rendered
differently.

On top of that, we removed the `utctime_` combinator and the stock
implementation for `HasDefaultColumnType`, because handling time is
largely complicated and needs to be addressed separately. For example,
see the extensive documentation <here
https://hackage.haskell.org/package/postgresql-simple-0.6.2/docs/Database-PostgreSQL-Simple-Time.html>.

Switching to support only `LocalTime` means we can restore the roundtrip
property between schemas.