Initial support for Bytea
This commit adds support for Bytea, which is treated as a
'binaryLargeObjectType' under the hood. For SqlSerial things are
slightly more complicated due to the fact there is no Pg.typoid for
'SERIAL', therefore beam-migrate opted for storing separate metadata
about this type for deserialisation purposes, which seems like what we
need to do here as well.