Skip sequence generation for PrimaryKeys
During the generic derivation code, we were generating a "companion
sequence" even for PrimaryKey(s), but this is incorrect as we want to
generate such sequence (and default constraint) only for the columnar
representing the field which will be referenced *as* the PrimaryKey.

While this still doesn't work 100% due to the fact we could embed the
same TableEntity as part of a DB, it's *almost* there.