Overhall Constraint/Sequence naming
Table constraints are now keyed according to what they constrain instead
of all of their attributes, including names.

No "type of all constraints, one at a time".  separate types for PRIMARY
KEY, UNIQUE, and FOREIGN KEY constraints; all split into the "match" and
"extra" parts.  All such constraints are now collected into a record of
maps, with the key being the match part of that constraint type and all
other data in the value of the map.

In most places for constraint/sequence, the name is wrapped in a Maybe,
to instruct beam-automigrate to allow postgres to choose the name, or
use the name from the db for existing, matching entity.

Sequences are now (optionally) owned by "SERIAL" columns using the built
in "OWNED BY" syntax in postgres.  A serial column can have an optional
name; beam-migrate will learn the name chosen by postgres at startup
when it already exists.

string parsing of DEFAULT constraints are now handled with a self
contained attoparsec based parser in `D.B.A.Parser`.  It's only parsing
the fragment of sql that beam-automigrate needs to understand to do it's
job; but is a good start for gradually improving its precision as the
need arises (future check-constraint support, as an example)

More liberal use of `"lens"` library to make some code a little easier
to understand.

Fixes dodgy catalog sql query for retrieving foreign key constraints.
how did this ever work?