Handle primary key migrations
Previous commits introduce a bug where no migration is assumed if both
versions of a table have a primary key at all,  This commit fixes the
case when both versions have a PK but with different columns.

This also exposed a related bug that primary key migrations must be
handled differently to unique key migrations.  a table can have only one
PK, which has further affects on when foreign key constraints must be
added or removed.